+++ internal/backend/mqttpubsub/base_test.go
@@ -3,7 +3,6 @@ package mqttpubsub
 import (
 	"os"
 
-	"github.com/brocaar/lora-gateway-bridge/internal/config"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -18,11 +17,6 @@ type conf struct {
 }
 
 func getConfig() *conf {
-	config.C.Backend.MQTT.DownlinkTopicTemplate = "gateway/{{ .MAC }}/tx"
-	config.C.Backend.MQTT.UplinkTopicTemplate = "gateway/{{ .MAC }}/rx"
-	config.C.Backend.MQTT.StatsTopicTemplate = "gateway/{{ .MAC }}/stats"
-	config.C.Backend.MQTT.AckTopicTemplate = "gateway/{{ .MAC }}/ack"
-
 	c := &conf{
 		Server: "tcp://127.0.0.1:1883",
 	}